Upgraded react-toastify
As mentioned in the upstream issue (https://gitlab.com/plum-tree/ui/-/issues/6), there were some problems with version 6 which prevented upgrading.

I'm not able to replicate those issues on version 9, but in case this turns out to break something, this is a separate commit so that it can be reverted easily later.