Re-added new react-toastify version, fixing the zooming / panning issue
The problem was due to event handlers being attached to the first svg element in the DOM, which happened to be an element inside the notification whenever one was open during page navigation.

Using an ID as a selector makes this much more robust now, as this uniquely identifies the correct element.