fix(database): Tree nodes no longer include all person data
This was redundantly saved and not kept up to date with the actual person data anyway.

To prevent weird data inconsistency issues, we now only save the ID.

To maintain compatibility with data saved before, we still keep an object around the ID, even though saving just the number would make more sense.