fix(import): Missing attributes are now replace with defaults if missing in a backup
The reason for this fix is a bug I'm unfortunately unable to reproduce.

Somehow I managed to delete the adoptiveParents and parentType attributes from the root node of a tree on mobile (or maybe it was missing from the start).

While this seems to not break anything else, importing such a broken tree fails when trying to map IDs. This should no longer be the case.